+++ feature-assets/src/main/java/io/novafoundation/nova/feature_assets/presentation/balance/list/BalanceListFragment.kt
@@ -139,7 +139,7 @@ class BalanceListFragment :
     }
 
     override fun tokenGroupClicked(tokenGroup: TokenGroupUi) {
-        showMessage("Not implemented yet")
+        viewModel.assetListMixin.expandToken(tokenGroup)
     }
 
     override fun totalBalanceClicked() {
